Refactor namespaces: create 'Model' namespace
Move types that model application configuration or state to the new
namespace XO.Console.Cli.Model. Types in this namespace may be public,
but should not be part of the core consumer-facing API.
