+++ XO.Console.Cli.Instrumentation/CommandAppInstrumentationOptions.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics;
+using XO.Console.Cli.Model;
 
 namespace XO.Console.Cli.Instrumentation;
