+++ XO.Console.Cli.Tests/CommandParametersBinderTest.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using XO.Console.Cli.Commands;
+using XO.Console.Cli.Model;
 using Xunit;
 
 namespace XO.Console.Cli.Tests;
