+++ XO.Console.Cli/Internal/Builtins.cs
@@ -1,4 +1,5 @@
 using System.Collections;
+using XO.Console.Cli.Model;
 
 namespace XO.Console.Cli;
