+++ XO.Console.Cli/Internal/CommandApp.cs
@@ -3,6 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using XO.Console.Cli.Commands;
+using XO.Console.Cli.Model;
 
 namespace XO.Console.Cli;
