+++ XO.Console.Cli/Internal/CommandBuilder.cs
@@ -1,6 +1,7 @@
 using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using XO.Console.Cli.Commands;
+using XO.Console.Cli.Model;
 
 namespace XO.Console.Cli;
