+++ XO.Console.Cli/Internal/MiddlewareAdapter.cs
@@ -1,3 +1,5 @@
+using XO.Console.Cli.Model;
+
 namespace XO.Console.Cli;
 
 internal sealed class MiddlewareAdapter
