+++ XO.Console.Cli/Model/CommandAppSettings.cs
@@ -1,6 +1,6 @@
 using System.Collections.Immutable;
 
-namespace XO.Console.Cli;
+namespace XO.Console.Cli.Model;
 
 internal sealed class CommandAppSettings
 {
