+++ XO.Console.Cli/Model/CommandTokenType.cs
@@ -1,4 +1,4 @@
-namespace XO.Console.Cli;
+﻿namespace XO.Console.Cli.Model;
 
 /// <summary>
 /// Describes the parser's interpretation of a command-line argument.
@@ -40,14 +40,3 @@ public enum CommandTokenType
     /// </summary>
     Unknown,
 }
-
-/// <summary>
-/// Represents the parser's interpretation of a command-line argument.
-/// </summary>
-/// <param name="TokenType">The type of token.</param>
-/// <param name="Value">The token value.</param>
-/// <param name="Context">An implementation-specific value describing how the token is to be used.</param>
-public sealed record CommandToken(
-    CommandTokenType TokenType,
-    string Value,
-    object? Context = null);
